avcodec/vc2enc: Initial vulkan VC2 encoder
Implements a Vulkan based dirac encoder. Supports Haar and Legall wavelets and should work with all wavelet depths.

Performance wise, encoding a 1080p 1-minute video is performed in about 2.5 minutes with the cpu encoder running on my Ryzen 5 4600H, while it takes about 30 seconds on my NVIDIA GTX 1650

Haar shader has a subgroup optimized variant that applies when configured wavelet depth allows it

lavapipe seems to be bugged for some reason, after a bunch of debugging I'm not quite sure if it's a bug here or in lavapipe. But people probably dont want to use this with a software implementation anyway.